wine_dataset_pytorch/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py38_1001
argon2-cffi=20.1.0=py38haf1e3a3_1
async_generator=1.10=py_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py38_0
cffi=1.14.3=py38hed5b41f_0
click=7.1.2=py_0
cycler=0.10.0=py_2
dbus=1.13.18=h18a8e69_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38_0
expat=2.2.10=hb1e8313_2
flask=1.1.2=py_0
freetype=2.10.4=h3f75d11_0
gettext=0.19.8.1=h15daf44_3
glib=2.66.1=h9bbe63b_0
icu=58.2=h0a44026_3
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
itsdangerous=1.1.0=py_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9d=hbcb3906_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py38h9250c2b_0
lcms2=2.11=h11f7e16_1
libblas=3.9.0=1_h508aa58_netlib
libcblas=3.9.0=3_h16b37ef_netlib
libcxx=11.0.0=h4c3b8ed_1
libffi=3.3=h74dc148_1
libgfortran=5.0.0=h7cc5361_13
libgfortran5=9.3.0=h7cc5361_13
libiconv=1.16=h1de35cc_0
liblapack=3.9.0=3_h16b37ef_netlib
libpng=1.6.37=h7cec526_2
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hca7d577_6
libuv=1.40.0=h22f3db7_0
libwebp-base=1.1.0=hbcb3906_3
llvm-openmp=11.0.0=h73239a0_1
lz4-c=1.9.2=hb1e8313_3
markupsafe=1.1.1=py38h1de35cc_1
matplotlib=3.3.3=py38h50d1736_0
matplotlib-base=3.3.3=py38h9bfd75d_0
mistune=0.8.4=py38h1de35cc_1001
mkl=2019.5=281
mkl-service=2.3.0=py38h0b31af3_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=h2e338ed_4
nest-asyncio=1.4.1=py_0
ninja=1.10.2=h9a9d8cb_0
notebook=6.1.4=py38_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1h=haf1e3a3_0
packaging=20.4=py_0
pandas=1.1.4=py38hcf432d8_0
pandoc=2.11=h0dc7051_0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=py38_1
pickleshare=0.7.5=py38_1001
pillow=8.0.1=py38h565d989_0
pip=20.3.1=pyhd8ed1ab_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
ptyprocess=0.6.0=py38_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.17.3=py38haf1e3a3_0
python=3.8.6=h3b7b5d6_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.7.0=py3.8_0
pytz=2020.4=pyhd8ed1ab_0
pyzmq=19.0.2=py38hb1e8313_1
qt=5.9.7=h468cd18_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h0678c8f_2
scikit-learn=0.23.2=py38h740a60c_3
scipy=1.5.3=py38hb6cd474_0
send2trash=1.5.0=py38_0
setuptools=49.6.0=py38h5347e94_2
sip=4.19.8=py38h0a44026_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.34.0=h17101e1_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_1
torchvision=0.8.1=py38_cpu
tornado=6.0.4=py38h1de35cc_1
tqdm=4.54.1=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=py_0
wheel=0.36.1=pyhd3deb0d_0
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=haf1e3a3_1
zeromq=4.3.3=hb1e8313_3
zipp=3.3.1=py_0
zlib=1.2.11=h7795811_1010
zstd=1.4.5=h289c70a_2
